+++ kafka/example/init_producer.go
@@ -44,7 +44,7 @@ func main() {
 			return []byte(req.ID)
 		}),
 		kafka.WithAcks(sarama.WaitForAll),
-		kafka.WithIdempotent(),
+		kafka.WithIdempotent(true),
 		kafka.WithRetryMax(3),
 		kafka.WithRetryBackoff(100*time.Millisecond),
 		kafka.WithCompression(sarama.CompressionSnappy),